feat : makeStick func 추가
