+++ CodeStarterCamp_Week1/icecream.swift
@@ -1,0 +1,25 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+
+import Foundation
+
+func makeIcecream() {
+    for _ in 1...8 {
+        print("***********")
+    }
+    
+}
+
+func makeStick() {
+    for _ in 1...4 {
+        print("    | |")
+    }
+}
+
+makeIcecream()
+makeStick()
